reorganize json-list and json-dict tests.
